Sync tests.

+++ afl/afl-test.c
@@ -55,7 +55,7 @@ main(int argc, char *argv[])
 		return(EXIT_FAILURE);
 	}
 
-	snprintf(buf, sizeof(buf), "%llu", st.st_size);
+	snprintf(buf, sizeof(buf), "%llu", (unsigned long long)st.st_size);
 	setenv("CONTENT_TYPE", "multipart/form-data; "
 		"boundary=---------------------------9051914041544843365972754266", 1);
 	setenv("REQUEST_METHOD", "post", 1);
